ticketbooking/requirements.txt
defusedxml==0.6.0
Django==2.0
django-allauth==0.40.0
django-environ==0.4.5
django-rest-auth==0.9.5
djangorestframework==3.10.2
oauthlib==3.1.0
python3-openid==3.1.0
requests-oauthlib==1.2.0
